graph_rag/core/generation.py
from transformers import AutoTokenizer, AutoModelForCausalLM
import torch

model_name = "TinyLlama/TinyLlama-1.1B-Chat-v1.0"  
tokenizer = AutoTokenizer.from_pretrained(model_name)
model = AutoModelForCausalLM.from_pretrained(model_name, torch_dtype=torch.float16, device_map="auto")

def generate_answer(question, contexts):
    context = "\n".join(contexts)
    prompt = f"""You are a financial analyst. Answer:
    1) Based on context mostly
    2) Without legal advice
    Context: {context}
    Question: {question}
    Answer (maximum 3 sentences):"""
    inputs = tokenizer(prompt, return_tensors="pt").to("cuda")
    outputs = model.generate(**inputs, max_new_tokens=100)
    return tokenizer.decode(outputs[0], skip_special_tokens=True)